bcrypt_pbkdf
Function bcrypt_pbkdf

pub fn bcrypt_pbkdf(
    passphrase: impl AsRef<[u8]>,
    salt: &[u8],
    rounds: u32,
    output: &mut [u8],
) -> Result<(), Error>

The bcrypt_pbkdf function.

Arguments

passphrase - The passphrase to process.

salt - The salt value to use as a byte vector.

rounds - The number of rounds to apply.

output - The resulting derived key is returned in this byte vector.

Returns

Ok(()) if everything is fine.

Err(Error::InvalidParamLen) if passphrase.is_empty() || salt.is_empty().

Err(Error::InvalidRounds) if rounds == 0.

Err(Error::InvalidOutputLen) if output.is_empty() || output.len() > 1024.